+++ vm/os_macos.cc
@@ -33,7 +33,7 @@ int64_t OS::CurrentMonotonicNanos() {
 const char* OS::Name() { return "macos"; }
 
 
-int OS::NumberOfAvailableProcessors() {
+intptr_t OS::NumberOfAvailableProcessors() {
   return sysconf(_SC_NPROCESSORS_ONLN);
 }
